Onboarding: rename stepper-config to step-counter-config
The file's scope is narrower than the original name suggested: it only
declares the logical-step groups consumed by the step counter.
