+++ client/landing/stepper/declarative-flow/flows/onboarding/step-counter-config.ts
